graphqlbackend: Move code into zoekt.go
textsearch.go is quite a large file, move most of the zoekt related code into
its own file.
